+++ docs/news.md
@@ -2,6 +2,12 @@
 
 This page documents notable changes introduced in each chronological release of the MOEA Framework.
 
+## Version 3.10 (TBD)
+
+  * Fixes bug #394 where changing the aggregate fitness comparator used by a GeneticAlgorithm would
+    not update the comparator used by its selection operator.
+
+
 ## Version 3.9 (05 Jan 2024)
 
   * Support BinaryIntegerVariable in externally-defined problems.  Namely, this sends the integer value